Separate members into separate JSON files
This has the advantage that many people can make PRs adding themselves
without causing merge conflicts.
