+++ src/members/Ezra-Crowe.json
@@ -1,0 +1,10 @@
+{
+  "name": "Ezra Crowe",
+  "pronouns": "he/him/",
+  "class": "'26",
+  "socials": [
+    "https://github.com/AbdominalBenito/obiesource.github.io",
+    "insta: ezra.crowe"
+  ],
+  "bio": "Im Ezra Crowe. Im an aspiring computer scientist from Seattle Washington. Im excited to be a part of Obie Source and start my journey of collaborative coding."
+}
