+++ src/members/_template.json
@@ -1,0 +1,22 @@
+{
+  "name": "Beans McLain",
+  "pronouns": "he/her/theirs",
+  "class": "'??",
+  "socials": [
+    "https://www.facebook.com/zuck",
+    "https://www.youtube.com/c/briandavidgilbert"
+  ],
+  "bio": "Hello! I'm a 6th year Geography major with a concentration in Computer Science. I like sticks and rocks and mud.",
+  "projects": [
+    {
+      "name": "google",
+      "description": "I made this.",
+      "website": "https://www.google.com"
+    },
+    {
+      "name": "Obiesource Website",
+      "description": "It's the Oberlin Open Source Website!",
+      "website": "https://obiesource.github.io"
+    }
+  ]
+}
