+++ src/members/mai.json
@@ -1,0 +1,7 @@
+{
+  "name": "Yen Mai",
+  "pronouns": "she/her/hers",
+  "class": "'24",
+  "socials": ["https://github.com/yen-mai-codes"],
+  "bio": "I am a 3rd year Computer Science major. I like manga and jewelry making."
+}
