+++ src/members/rehwinkel.json
@@ -1,0 +1,22 @@
+{
+  "name": "William Rehwinkel",
+  "pronouns": "he/him/his",
+  "class": "'24",
+  "socials": [
+    "https://koyu.space/@fiskfan1999",
+    "https://spectra.video/c/willrehwinkel_organ/videos"
+  ],
+  "bio": "I'm a 3rd year double-degree student majoring in Computer Science and the Pipe Organ currently living in the New York/New Jersey area. I particularly love open-source projects, not only using them, but contributing to them and developing my own. My favorites are ffmpeg, ergo (an IRC daemon, which I'm particularly proud of having contributed to), and the Tor network. Recently, I moved to a new peertube instance which you can find above.",
+  "projects": [
+    {
+      "name": "FiskFan1999/wvlist",
+      "description": "Catalogue of composers list of works with community submissions",
+      "website": "https://wvlist.net"
+    },
+    {
+      "name": "ObieSource/peertube-multipart-upload",
+      "description": "Command-line application to upload videos to a Peertube instance using the API.",
+      "website": "https://github.com/ObieSource/peertube-multipart-upload"
+    }
+  ]
+}
