+++ src/members/snyder.json
@@ -1,0 +1,7 @@
+{
+  "name": "Ben Snyder",
+  "pronouns": "he/him",
+  "class": "'25",
+  "socials": ["https://www.instagram.com/bensnydoor"],
+  "bio": "I'm a 2nd year CS major"
+}
