+++ src/members/toker.json
@@ -1,0 +1,14 @@
+{
+  "name": "Ben Toker",
+  "pronouns": "he/him/his",
+  "class": "Class of 2026",
+  "socials": ["https://instagram.com/tenboker"],
+  "bio": "Possibly cool.",
+  "projects": [
+    {
+      "name": "webScrape",
+      "description": "I made this.",
+      "website": "https://btoker.dev"
+    }
+  ]
+}
